move switches to switches.py
